+++ packages/components/panel/index.ts
@@ -1,0 +1,9 @@
+import Panel from './src/Panel.vue'
+import { withInstall } from '@zl-gp/utils'
+
+export const ZlPanel = withInstall(Panel)
+
+export * from './src/instances'
+export * from './src/types'
+
+export default ZlPanel