+++ packages/components/panel/src/instances.ts
@@ -1,0 +1,3 @@
+import Panel from './Panel.vue'
+
+export type PanelInstance = InstanceType<typeof Panel>