+++ packages/components/panel/src/types.ts
@@ -1,0 +1,6 @@
+export interface PanelProps {
+  title: string
+  name: string
+  forTab: boolean
+  isFolder: boolean
+}