+++ packages/components/tabs/src/instances.ts
@@ -1,0 +1,5 @@
+import Tabs from './Tabs.vue'
+import TabPanel from './TabPanel.vue'
+export type TabsInstance = InstanceType<typeof Tabs>
+
+export type TabPanelInstance = InstanceType<typeof TabPanel>