+++ packages/play/src/pages/TabDemo.vue
@@ -1,0 +1,15 @@
+<script setup lang="ts"></script>
+<template>
+  <zl-tabs>
+    <zl-tab-panel name="hi" title="xx" :for-tab="true">sss</zl-tab-panel>
+    <zl-tab-panel name="hi1" title="xx1" :for-tab="true">
+      <p>qsaa</p>
+      <p>qsaa</p>
+      <p>qsaa</p>
+      <p>qsaa</p>
+      <p>qsaa</p>
+      <p>qsaa</p>
+      <p>qsaa</p>
+    </zl-tab-panel>
+  </zl-tabs>
+</template>